# api/openapi.yaml
openapi: 3.1.0
info:
  title: OneImg API
  version: 1.0.0
  description: Breaking v1 API for OneImg. Legacy /api routes are not supported.
servers:
  - url: /api/v1
...
paths:
...
  /images:
    get:
      operationId: listImages
      tags: [Images]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/Order' }
        - { name: q, in: query, schema: { type: string } }
        - { name: bucket_id, in: query, schema: { type: integer, minimum: 1 } }
        - { name: tag_ids, in: query, description: Comma-separated positive IDs, schema: { type: string } }
        - { name: untagged, in: query, schema: { type: boolean } }
        - { name: uploader_role, in: query, schema: { type: integer, enum: [1, 3] } }
      responses:
        '200': { $ref: '#/components/responses/ImageCollectionResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    post:
      operationId: uploadImages
      tags: [Images]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images: { type: array, minItems: 1, maxItems: 10, items: { type: string, format: binary } }
                tag_ids: { type: array, items: { type: integer, minimum: 1 } }
                bucket_id: { type: integer, minimum: 1 }
      responses:
        '200': { $ref: '#/components/responses/UploadBatchResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /images/{id}:
    parameters: [{ $ref: '#/components/parameters/ID' }]
    get:
      operationId: getImage
      tags: [Images]
      responses:
        '200': { $ref: '#/components/responses/ImageResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
    delete:
      operationId: deleteImage
      tags: [Images]
      responses:
        '204': { $ref: '#/components/responses/NoContentResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
  /image-imports:
    post:
      operationId: importImage
      tags: [Images]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ImageImportRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/ImageResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
  /images/tags:
    patch:
      operationId: updateImageTags
      tags: [Images]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BatchImageTagsRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/DataResponse' }
        default: { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: oneimg-session }
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, schema: { type: string } }
    Order: { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
...
  responses:
    ProblemResponse:
      description: RFC 9457 Problem Details
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    DataResponse:
      description: Successful response
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/DataEnvelope' } } }
    NoContentResponse:
      description: Successful response with no content
      headers: { X-Request-ID: { schema: { type: string } } }
...
    ImageResponse:
      description: Image response
      headers: { X-Request-ID: { schema: { type: string } }, Location: { schema: { type: string } } }
      content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/Image' } } } } }
    ImageCollectionResponse:
      description: Paginated image response
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { type: object, required: [data, meta], properties: { data: { type: array, items: { $ref: '#/components/schemas/Image' } }, meta: { $ref: '#/components/schemas/Meta' } } } } }
...
    UploadBatchResponse:
      description: Per-file upload results
      headers: { X-Request-ID: { schema: { type: string } } }
      content: { application/json: { schema: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/UploadBatch' } } } } }
...
  schemas:
    DataEnvelope:
      type: object
      required: [data]
      properties: { data: {}, meta: { $ref: '#/components/schemas/Meta' } }
    Meta:
      type: object
      properties:
        pagination:
          type: object
          required: [page, page_size, total, total_pages]
          properties: { page: { type: integer }, page_size: { type: integer }, total: { type: integer, format: int64 }, total_pages: { type: integer, format: int64 } }
    FieldProblem:
      type: object
      required: [field, code, message]
      properties: { field: { type: string }, code: { type: string }, message: { type: string } }
    Problem:
      type: object
      required: [type, title, status, detail, code, instance, request_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors: { type: array, items: { $ref: '#/components/schemas/FieldProblem' } }
...
    Tag:
      type: object
      required: [id, name]
      properties: { id: { type: integer }, name: { type: string } }
    Image:
      type: object
      required: [id, url, thumbnail, file_name, original_file_name, original_file_size, file_size, mime_type, width, height, storage, bucket_id, user_id, uploader_role, tags, created_at]
      properties:
        id: { type: integer }
        url: { type: string }
        thumbnail: { type: string }
        file_name: { type: string }
        original_file_name: { type: string }
        original_file_size: { type: integer, format: int64 }
        file_size: { type: integer, format: int64 }
        mime_type: { type: string }
        width: { type: integer }
        height: { type: integer }
        storage: { type: string }
        bucket_id: { type: integer }
        user_id: { type: integer }
        uploader_role: { type: integer }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
        created_at: { type: string, format: date-time }
...
    UploadFileResult:
      type: object
      required: [file_name, success]
      properties:
        file_name: { type: string }
        success: { type: boolean }
        duplicate: { type: boolean }
        image: { $ref: '#/components/schemas/Image' }
        error:
          type: object
          required: [code, detail]
          properties: { code: { type: string }, detail: { type: string } }
    UploadBatch:
      type: object
      required: [files, summary]
      properties:
        files: { type: array, items: { $ref: '#/components/schemas/UploadFileResult' } }
        summary:
          type: object
          required: [total, succeeded, failed]
          properties:
            total: { type: integer }
            succeeded: { type: integer }
            failed: { type: integer }
...
    ImageImportRequest:
      type: object
      required: [url]
      properties: { url: { type: string, format: uri }, bucket_id: { type: integer, minimum: 1 }, tag_ids: { type: array, items: { type: integer, minimum: 1 } } }
    BatchImageTagsRequest:
      type: object
      required: [image_ids]
      properties: { image_ids: { type: array, minItems: 1, items: { type: integer, minimum: 1 } }, add_tag_ids: { type: array, items: { type: integer, minimum: 1 } }, remove_tag_ids: { type: array, items: { type: integer, minimum: 1 } } }
...
security:
  - cookieAuth: []
  - bearerAuth: []